requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.4
idna==2.10
joblib==0.16.0
multidict==4.7.6
nltk==3.5
numpy==1.19.1
regex==2020.7.14
scikit-learn==0.23.1
scipy==1.5.2
threadpoolctl==2.1.0
tqdm==4.48.0
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.5.0
